Cruds restantes

+++ frontend/src/app/pages/employee/employee.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Author } from 'src/app/models/author';
 import { Book } from 'src/app/models/book';
+import { AuthorService } from 'src/app/services/author.service';
 import { BookService } from 'src/app/services/book.service';
 
 @Component({
@@ -12,28 +14,30 @@ export class EmployeeComponent implements OnInit {
   pageSize = 10;
 
   public books: Array<Book> = [];
+  public authors: Array<Author> = [];
 
-  constructor(private bookService: BookService) { }
+  constructor(private bookService: BookService, private authorService: AuthorService) { }
 
   ngOnInit() {
-    //console.log("hola holita")
-    this.loadInfo();
+    
     if(localStorage.getItem('rol') != 'EMPLOYEE'){
       window.location.assign('/');
     }
+
+    this.loadInfo();
   }
 
   loadInfo() {
     this.bookService.getBooks().subscribe((response) => {
-      //console.log("load libro antes")
       this.books = response;
-      //console.log("load libro despues")
+    })
+
+    this.authorService.getAuthors().subscribe((response) => {
+      this.authors = response;
     })
   }
 
-updateLibro(book: Book, name: string, author: string, genre: string, synopsis: string, stock: number, image: string, fk: string, e) {
-    //const 22auxFk = Number(fk);
-    //console.log("Llegó aupdate libro")
+  updateLibro(book: Book, name: string, author: string, genre: string, synopsis: string, stock: number, image: string, fk: string, e) {
     
     e.preventDefault();
     const newBook: Book = {
@@ -50,16 +54,47 @@ updateLibro(book: Book, name: string, author: string, genre: string, synopsis: s
     console.log(newBook);
 
     this.bookService.updateLibro(newBook.id, newBook).subscribe(() => {
-      // console.log("update libro antes")
       this.loadInfo();
-      // console.log("update libro despues")
+    });
+
+  }
+
+  updateAutor(author: Author, name: string, e){
+    
+    e.preventDefault();
+    const newAuthor: Author = {
+      id: author.id,
+      name: name
+    }
+
+    console.log(newAuthor);
+
+    this.authorService.updateAuthor(newAuthor.id, newAuthor).subscribe(() => {
+      this.loadInfo();
     });
 
   }
 
   deleteLibro(id: number) {
-    this.bookService.deleteLibro(id).subscribe(() => {
-      this.loadInfo();
-    })
+    if(confirm("Pulse aceptar para borrar el libro") == true){
+      this.bookService.deleteLibro(id).subscribe(() => {
+        this.loadInfo();
+      })
+    }else{
+
+    }
+    
   }
+
+  deleteAuthor(id: number) {
+    if(confirm("Pulse aceptar para borrar el autor") == true){
+      this.authorService.deleteAuthor(id).subscribe(() => {
+        this.loadInfo();
+      })
+    }else{
+
+    }
+    
+  }
+
 }
